ts json-ast: add AST structs
